requirements-tests.txt
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==6.2.2;python_version<"3.0"
Pillow==7.2.0;python_version>="3.0"
PyYAML==5.3.1
Shapely==1.7.0
WebOb==1.8.6
WebTest==2.0.35
attrs==19.3.0
aws-sam-translator==1.26.0
aws-xray-sdk==2.6.0
beautifulsoup4==4.9.1
boto3==1.14.46
azure-storage-blob==12.8.0
boto==2.49.0
botocore==1.17.46
certifi==2020.6.20
cffi==1.14.2
cfn-lint==0.35.0
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
docker==4.3.0
docutils==0.15.2
ecdsa==0.14.1
future==0.18.2
idna==2.8
importlib-metadata==1.7.0
iniconfig==1.0.1
jmespath==0.10.0
jsondiff==1.1.2
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
lxml==4.5.2
mock==3.0.5;python_version<"3.6"
mock==4.0.2;python_version>="3.6"
more-itertools==5.0.0;python_version<"3.0"
more-itertools==8.4.0;python_version>="3.0"
moto==1.3.14
networkx==2.2;python_version<"3.0"
networkx==2.4;python_version>="3.0"
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.2.2;python_version<"3.0"
pyparsing==2.4.7;python_version>="3.0"
pyproj==2.2.2;python_version<"3.0"
pyproj==2.6.1.post1;python_version>="3.0"
pyrsistent==0.16.0
pytest==4.6.11;python_version<"3.0"
pytest==6.0.1;python_version>="3.0"
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
responses==0.10.16
riak==2.7.0
rsa==4.5;python_version<"3.0"
rsa==4.6;python_version>="3.0"
s3transfer==0.3.3
six==1.15.0
soupsieve==1.9.6;python_version<"3.0"
soupsieve==2.0.1;python_version>="3.0"
sshpubkeys==3.1.0
toml==0.10.1
urllib3==1.25.10
waitress==1.4.4
websocket-client==0.57.0
werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==1.2.0;python_version<"3.6"
zipp==3.1.0;python_version>="3.6"
sphinx_bootstrap_theme==0.7.1

